rfs: Fix for 64-bit targets
The RTEMS_BLKIO_SETBLKSIZE IO control expects an uint32_t parameter and
not a size_t which is 64-bits on 64-bit targets.

Update #3082.

+++ cpukit/libfs/src/rfs/rtems-rfs-buffer.c
@@ -411,7 +411,7 @@ rtems_rfs_buffer_sync (rtems_rfs_file_system* fs)
 }
 
 int
-rtems_rfs_buffer_setblksize (rtems_rfs_file_system* fs, size_t size)
+rtems_rfs_buffer_setblksize (rtems_rfs_file_system* fs, uint32_t size)
 {
   int rc;
